Renamed `logging` --> `logger` to avoid inbuilt conflict
