+++ minerva/logger/__init__.py
@@ -23,7 +23,7 @@
 #
 # @org: University of Southampton
 # Created under a project funded by the Ordnance Survey Ltd.
-""":mod:`logging` contains functionality for logging and analysing results within :mod:`minerva`"""
+""":mod:`minerva.logger` contains functionality for logging and analysing results within :mod:`minerva`"""
 # =====================================================================================================================
 #                                                    METADATA
 # =====================================================================================================================
